switching dimensions to int64_t to match cudensitymat

+++ runtime/cudaq/dynamics/boson_operators.h
@@ -37,7 +37,7 @@ private:
 
   std::string op_code_to_string() const;
   virtual std::string
-  op_code_to_string(std::unordered_map<std::size_t, int> &dimensions) const override;
+  op_code_to_string(std::unordered_map<std::size_t, int64_t> &dimensions) const override;
 
   void inplace_mult(const boson_handler &other);
 
@@ -61,7 +61,7 @@ public:
   /// @arg  `dimensions` : A map specifying the dimension, that is the number of
   /// eigenstates, for each degree of freedom.
   virtual complex_matrix
-  to_matrix(std::unordered_map<std::size_t, int> &dimensions,
+  to_matrix(std::unordered_map<std::size_t, int64_t> &dimensions,
             const std::unordered_map<std::string, std::complex<double>>
                 &parameters = {}) const override;
