+++ runtime/cudaq/dynamics/helpers.h
@@ -31,7 +31,7 @@ using EigenSparseMatrix = Eigen::SparseMatrix<std::complex<double>, 0x1, long>; 
 /// to the sequence of degrees (ordering is relevant if dimensions differ).
 std::vector<std::string>
 generate_all_states(const std::vector<std::size_t> &degrees,
-                    const std::unordered_map<std::size_t, int> &dimensions);
+                    const std::unordered_map<std::size_t, int64_t> &dimensions);
 
 /// Computes a vector describing the permutation to reorder a matrix that is
 /// ordered according to `op_degrees` to apply to `canon_degrees` instead.
@@ -40,7 +40,7 @@ generate_all_states(const std::vector<std::size_t> &degrees,
 std::vector<std::size_t>
 compute_permutation(const std::vector<std::size_t> &op_degrees,
                     const std::vector<std::size_t> &canon_degrees,
-                    const std::unordered_map<std::size_t, int> dimensions);
+                    const std::unordered_map<std::size_t, int64_t> dimensions);
 
 /// Permutes the given matrix according to the given permutation.
 /// If states is the current order of vector entries on which the given matrix
