+++ runtime/cudaq/dynamics/matrix_operators.cpp
@@ -49,7 +49,7 @@ std::string matrix_handler::type_prefix<fermion_handler>() {
 }
 
 void matrix_handler::define(std::string operator_id,
-                            std::vector<int> expected_dimensions,
+                            std::vector<int64_t> expected_dimensions,
                             matrix_callback &&create) {
   auto defn = Definition(operator_id, expected_dimensions,
                          std::forward<matrix_callback>(create));
@@ -109,7 +109,7 @@ matrix_handler::instantiate(std::string operator_id, std::vector<std::size_t> &&
 // private helpers
 
 std::string matrix_handler::op_code_to_string(
-    std::unordered_map<std::size_t, int> &dimensions) const {
+    std::unordered_map<std::size_t, int64_t> &dimensions) const {
   auto it = matrix_handler::defined_ops.find(this->op_code);
   assert(it != matrix_handler::defined_ops
                    .end()); // should be validated upon instantiation
@@ -154,7 +154,7 @@ matrix_handler::matrix_handler(std::size_t degree)
   if (matrix_handler::defined_ops.find(this->op_code) ==
       matrix_handler::defined_ops.end()) {
     auto func =
-        [](const std::vector<int> &dimensions,
+        [](const std::vector<int64_t> &dimensions,
            const std::unordered_map<std::string, std::complex<double>> &_none) {
           std::size_t dimension = dimensions[0];
           auto mat = complex_matrix(dimension, dimension);
@@ -231,10 +231,10 @@ matrix_handler::matrix_handler(const T &other,
       targets(other.degrees()) {
   if (matrix_handler::defined_ops.find(this->op_code) ==
       matrix_handler::defined_ops.end()) {
-    auto func = [other](const std::vector<int> &dimensions,
+    auto func = [other](const std::vector<int64_t> &dimensions,
                         const std::unordered_map<std::string,
                                                  std::complex<double>> &_none) {
-      std::unordered_map<std::size_t, int> dims;
+      std::unordered_map<std::size_t, int64_t> dims;
       auto targets = other.degrees();
       for (auto i = 0; i < dimensions.size(); ++i)
         dims[targets[i]] = dimensions[i];
@@ -242,7 +242,7 @@ matrix_handler::matrix_handler(const T &other,
     };
     // the to_matrix method on the spin op will check the dimensions, so we
     // allow arbitrary here
-    std::vector<int> required_dimensions(this->targets.size(), -1);
+    std::vector<int64_t> required_dimensions(this->targets.size(), -1);
     matrix_handler::define(this->op_code, std::move(required_dimensions), func);
   }
 }
@@ -305,14 +305,14 @@ matrix_handler::operator=(const fermion_handler &other);
 // evaluations
 
 complex_matrix matrix_handler::to_matrix(
-    std::unordered_map<std::size_t, int> &dimensions,
+    std::unordered_map<std::size_t, int64_t> &dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters)
     const {
   auto it = matrix_handler::defined_ops.find(this->op_code);
   assert(it != matrix_handler::defined_ops
                    .end()); // should be validated upon instantiation
 
-  std::vector<int> relevant_dimensions;
+  std::vector<int64_t> relevant_dimensions;
   relevant_dimensions.reserve(this->targets.size());
   for (auto i = 0; i < this->targets.size(); ++i) {
     auto entry = dimensions.find(this->targets[i]);
@@ -364,7 +364,7 @@ product_op<matrix_handler> matrix_handler::number(std::size_t degree) {
   if (matrix_handler::defined_ops.find(op_code) ==
       matrix_handler::defined_ops.end()) {
     auto func =
-        [](const std::vector<int> &dimensions,
+        [](const std::vector<int64_t> &dimensions,
            const std::unordered_map<std::string, std::complex<double>> &_none) {
           std::size_t dimension = dimensions[0];
           auto mat = complex_matrix(dimension, dimension);
@@ -384,7 +384,7 @@ product_op<matrix_handler> matrix_handler::parity(std::size_t degree) {
   if (matrix_handler::defined_ops.find(op_code) ==
       matrix_handler::defined_ops.end()) {
     auto func =
-        [](const std::vector<int> &dimensions,
+        [](const std::vector<int64_t> &dimensions,
            const std::unordered_map<std::string, std::complex<double>> &_none) {
           std::size_t dimension = dimensions[0];
           auto mat = complex_matrix(dimension, dimension);
@@ -404,7 +404,7 @@ product_op<matrix_handler> matrix_handler::position(std::size_t degree) {
   if (matrix_handler::defined_ops.find(op_code) ==
       matrix_handler::defined_ops.end()) {
     auto func =
-        [](const std::vector<int> &dimensions,
+        [](const std::vector<int64_t> &dimensions,
            const std::unordered_map<std::string, std::complex<double>> &_none) {
           std::size_t dimension = dimensions[0];
           auto mat = complex_matrix(dimension, dimension);
@@ -426,7 +426,7 @@ product_op<matrix_handler> matrix_handler::momentum(std::size_t degree) {
   if (matrix_handler::defined_ops.find(op_code) ==
       matrix_handler::defined_ops.end()) {
     auto func =
-        [](const std::vector<int> &dimensions,
+        [](const std::vector<int64_t> &dimensions,
            const std::unordered_map<std::string, std::complex<double>> &_none) {
           std::size_t dimension = dimensions[0];
           auto mat = complex_matrix(dimension, dimension);
@@ -449,7 +449,7 @@ product_op<matrix_handler> matrix_handler::displace(std::size_t degree) {
   std::string op_code = "displace";
   if (matrix_handler::defined_ops.find(op_code) ==
       matrix_handler::defined_ops.end()) {
-    auto func = [](const std::vector<int> &dimensions,
+    auto func = [](const std::vector<int64_t> &dimensions,
                    const std::unordered_map<std::string, std::complex<double>>
                        &parameters) {
       std::size_t dimension = dimensions[0];
@@ -477,7 +477,7 @@ product_op<matrix_handler> matrix_handler::squeeze(std::size_t degree) {
   std::string op_code = "squeeze";
   if (matrix_handler::defined_ops.find(op_code) ==
       matrix_handler::defined_ops.end()) {
-    auto func = [](const std::vector<int> &dimensions,
+    auto func = [](const std::vector<int64_t> &dimensions,
                    const std::unordered_map<std::string, std::complex<double>>
                        &parameters) {
       std::size_t dimension = dimensions[0];
