+++ runtime/cudaq/dynamics/product_operators.cpp
@@ -635,7 +635,7 @@ std::string product_op<HandlerTy>::to_string() const {
 
 template <typename HandlerTy>
 complex_matrix product_op<HandlerTy>::to_matrix(
-    std::unordered_map<std::size_t, int> dimensions,
+    std::unordered_map<std::size_t, int64_t> dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters,
     bool application_order) const {
   auto evaluated =
@@ -657,7 +657,7 @@ complex_matrix product_op<HandlerTy>::to_matrix(
 
 template <>
 complex_matrix product_op<spin_handler>::to_matrix(
-    std::unordered_map<std::size_t, int> dimensions,
+    std::unordered_map<std::size_t, int64_t> dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters,
     bool application_order) const {
   auto terms = std::move(
@@ -679,7 +679,7 @@ complex_matrix product_op<spin_handler>::to_matrix(
   template std::string product_op<HandlerTy>::to_string() const;               \
                                                                                \
   template complex_matrix product_op<HandlerTy>::to_matrix(                    \
-      std::unordered_map<std::size_t, int> dimensions,                         \
+      std::unordered_map<std::size_t, int64_t> dimensions,                     \
       const std::unordered_map<std::string, std::complex<double>> &parameters, \
       bool application_order) const;
 
@@ -1293,7 +1293,7 @@ std::size_t product_op<HandlerTy>::num_qubits() const {
 HANDLER_SPECIFIC_TEMPLATE_DEFINITION(spin_handler)
 std::string
 product_op<HandlerTy>::get_pauli_word(std::size_t pad_identities) const {
-  std::unordered_map<std::size_t, int> dims;
+  std::unordered_map<std::size_t, int64_t> dims;
   auto terms = std::move(
       this->evaluate(
               operator_arithmetics<operator_handler::canonical_evaluation>(dims,
@@ -1338,7 +1338,7 @@ std::vector<bool> product_op<HandlerTy>::get_binary_symplectic_form() const {
   if (this->operators.size() == 0)
     return {};
 
-  std::unordered_map<std::size_t, int> dims;
+  std::unordered_map<std::size_t, int64_t> dims;
   auto degrees = this->degrees(
       false); // degrees in canonical order to match the evaluation
   auto evaluated = this->evaluate(
@@ -1372,7 +1372,7 @@ std::vector<bool> product_op<HandlerTy>::get_binary_symplectic_form() const {
 
 HANDLER_SPECIFIC_TEMPLATE_DEFINITION(spin_handler)
 csr_spmatrix product_op<HandlerTy>::to_sparse_matrix(
-    std::unordered_map<std::size_t, int> dimensions,
+    std::unordered_map<std::size_t, int64_t> dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters,
     bool application_order) const {
   auto terms = std::move(
@@ -1395,7 +1395,7 @@ product_op<spin_handler>::get_pauli_word(std::size_t pad_identities) const;
 template std::vector<bool>
 product_op<spin_handler>::get_binary_symplectic_form() const;
 template csr_spmatrix product_op<spin_handler>::to_sparse_matrix(
-    std::unordered_map<std::size_t, int> dimensions,
+    std::unordered_map<std::size_t, int64_t> dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters,
     bool application_order) const;
